chore: update rules for release service catalog testing
