+++ magefiles/rulesengine/engine/engine.go
@@ -9,6 +9,7 @@ var MageEngine = rulesengine.RuleEngine{
 	"tests": {
 		"e2e-repo":          repos.E2ETestRulesCatalog,
 		"infra-deployments": repos.InfraDeploymentsTestRulesCatalog,
+		"release-catalog":   repos.ReleaseServiceCatalogTestRulesCatalog,
 	},
 	"demo": {
 		"local-workflow": repos.DemoCatalog,
