+++ magefiles/rulesengine/repos/e2e_repo.go
@@ -2,7 +2,7 @@ package repos
 
 import (
 	"strings"
-	"strconv"
+
 	"github.com/konflux-ci/e2e-tests/magefiles/rulesengine"
 )
 
@@ -22,45 +22,7 @@ var TestFilesOnlyRule = rulesengine.Rule{Name: "E2E PR Test File Diff Execution"
 		rulesengine.ConditionFunc(CheckNoFilesChanged)},
 	Actions: []rulesengine.Action{rulesengine.ActionFunc(ExecuteFocusedFileAction)}}
 
-var ReleaseCatalogPairedRule = rulesengine.Rule{Name: "Release Catalog PR paired Test Execution",
-	Description: "Runs release catalog tests except for the fbc tests on release-service-catalog repo when PR paired and not a rehearsal job",
-	Condition: rulesengine.All{rulesengine.ConditionFunc(releaseCatalogRepoCondition),
-		rulesengine.ConditionFunc(isPaired),
-		rulesengine.None{
-			rulesengine.ConditionFunc(isRehearse),
-		},
-	},
-	Actions: []rulesengine.Action{rulesengine.ActionFunc(ExecuteReleasePairedAction)}}
-
-var ReleaseServiceCatalogRule = rulesengine.Rule{Name: "Release Service Catalog Test Execution",
-	Description: "Runs all release catalog tests on release-service-catalog repo on PR/rehearsal jobs",
-	Condition: rulesengine.All{rulesengine.ConditionFunc(releaseCatalogRepoCondition),
-		rulesengine.Any{
-			rulesengine.None{rulesengine.ConditionFunc(isPaired),},
-			rulesengine.ConditionFunc(isRehearse),
-		},
-	},
-	Actions: []rulesengine.Action{rulesengine.ActionFunc(ExecuteReleaseCatalogAction)}}
-
-var E2ETestRulesCatalog = rulesengine.RuleCatalog{NonTestFilesRule, TestFilesOnlyRule, ReleaseServiceCatalogRule, ReleaseCatalogPairedRule}
-
-var isRehearse = func(rctx *rulesengine.RuleCtx) bool {
-
-	return strings.Contains(rctx.JobName, "rehearse")
-}
-
-var isPaired = func(rctx *rulesengine.RuleCtx) bool {
-
-	if boolValue, err := strconv.ParseBool(rctx.IsPaired); err == nil && boolValue {
-		return true
-	}
-	return false
-}
-
-func releaseCatalogRepoCondition(rctx *rulesengine.RuleCtx) bool {
-
-	return rctx.RepoName == "release-service-catalog"
-}
+var E2ETestRulesCatalog = rulesengine.RuleCatalog{NonTestFilesRule, TestFilesOnlyRule}
 
 func CheckNoFilesChanged(rctx *rulesengine.RuleCtx) bool {
 
@@ -131,14 +93,4 @@ func ExecuteFocusedFileAction(rctx *rulesengine.RuleCtx) error {
 
 	return ExecuteTestAction(rctx)
 
-}
-
-func ExecuteReleasePairedAction(rctx *rulesengine.RuleCtx) error {
-	rctx.LabelFilter = "release-pipelines && !fbc-tests"
-	return ExecuteTestAction(rctx)
-}
-
-func ExecuteReleaseCatalogAction(rctx *rulesengine.RuleCtx) error {
-	rctx.LabelFilter = "release-pipelines"
-	return ExecuteTestAction(rctx)
 }
